Modified _get_localzone() to use _root param
The _unix._get_localzone() function takes an optional root path to use
when searching for timezone files which is intended for use in tests.
This root path was added to the search location for 'etc/localtime',
which allows for the removal of the `pyfakefs` dependency from the unit
tests. However, this requires the unit tests to call _get_localzone()
directly, not its public wrapper function.